A different approach to Cook Mode styling
I’ve added a cards theme, which makes card pages reusable, and moved
Cook Mode styling out of application.css. There’s a lot of very
different styling for CM, so not loading it on all the pages makes a
lot of sense. And of course, the markup has been reduced quite a lot.