rewrite save
We’re now using one standard way to save recipes off the Internet, via
the SaveController and the ScrapingHelper. The clipper uses the new
way, and the future save buttons will use it. The scraping itself also
relies more on Hangry and less on my brittle, handmade scrapers, which
allows support for thousands of sites.